requirements.txt
autopep8==1.4.4
pep8==1.7.1
pycodestyle==2.5.0
sqlparse==0.2.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2025.8.3
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.2
debugpy==1.8.6
decorator==5.1.1
dill==0.3.8
Django==5.1.7
django-allauth==65.7.0
django-environ==0.12.0
django-filter==25.1
django-qr-code==4.2.0
djangorestframework==3.16.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlip==0.2
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.2
RapidFuzz==3.13.0
reportlab==4.4.4
requests==2.32.5
segno==1.6.6
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.16.0